move fetch of CTR report up a component
